Envelope System Overhaul Pt.27
It appears that I counted Pt.12 twice... oh well.

Removed many envelopes to shorten the envelope list as much as I could. Added copy/paste and reordering buttons to envelopes. Tweaked envelope importing and removed envelope variants. Did a few more mod capitalization checks.

Now, what still needs to be done is to restructure the UI of the copy/paste and move up/down buttons, and to fix drum channels and the drumset.