Envelope System Overhaul Pt.32
Functional Build!

Finalized the drumset's envelope UI, added various new buttons for copy/pasting, reordering, randomizing, and adding copied envelopes, as well as one more button which automatically sets the pitch envelope's start/end points with the lowest/highest notes in a selection. Added a new preference which toggles the visibility of reordering buttons and completely changed the UI layout of envelopes. Optimized some code.

It's time to move on to some more amazing features with this mini-project!