Envelope System Overhaul Pt.38
DEFUNCTIONAL BUILD

More progress on the basic custom envelope, this time restructuring the synth to be more aligned with some better future plans on what to do with it. Added a new usefulStuff function called "bounce mod" which acts like a modulo with an additional bound that makes wavefold-type shapes. Added a new LFO shape called "curve tooth" and rewrited LFO shapes into an enum.

Progress on the build was halted due to many undefined and song corruption issues stacking up, need help.